feat(399): 신청 폼 주소 직접입력 기능 추가, 페이지 정리
