+++ components/request/modals/index.ts
@@ -1,0 +1,7 @@
+import AlertModal from "./AlertModal";
+import ConfirmModal from "./ConfirmModal";
+import DuplicatedModal from "./DuplicatedModal";
+import SubmitModal from "./SubmitModal";
+
+export { AlertModal, ConfirmModal, DuplicatedModal, SubmitModal };
+export default {};
