+++ components/request/place/PlaceInput.tsx
@@ -1,15 +1,15 @@
 /* eslint-disable  @typescript-eslint/no-explicit-any */
-import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { FaSearch, FaTimes } from "react-icons/fa";
 import { useRecoilState } from "recoil";
 import { requestInputsAtom } from "../../../shared/state";
+import BasicInput from "../units/BasicInput";
 import SearchInput from "../units/SearchInput";
 import {
 	StyledSearchListContainer,
 	StyledSearchList,
 } from "../units/searchListStyle";
-import { StyledPlaceInput } from "./placeInputStyle";
+import { StyledPlaceCustomInput, StyledPlaceInput } from "./placeInputStyle";
 
 declare global {
 	interface Window {
@@ -23,23 +23,19 @@ type KakaoResult = {
 	road_address_name: string; // "서울 용산구 보광로 126"
 };
 
-const KakaoAxios = axios.create({
-	baseURL: "//dapi.kakao.com",
-	headers: {
-		Authorization: `KakaoAK ${process.env.NEXT_PUBLIC_KAKAO_KEY}`,
-	},
-});
-
-// 법정동 코드 검색
-const bCodeSearch = (params: { query: string }) =>
-	KakaoAxios.get("/v2/local/search/address.json", { params });
-
 const PlaceInput = () => {
 	const [requestInputs, setRequestInputs] = useRecoilState(requestInputsAtom);
 
 	const [isSearchOpen, setSearchOpen] = useState(false);
 	const [keyword, setKeyword] = useState("");
 	const [placeList, setPlaceList] = useState([] as KakaoResult[]);
+
+	const [isInputOpen, setInputOpen] = useState(false);
+	const [customPlace, setCustomPlace] = useState({
+		place: "",
+		address1: "",
+		address2: "",
+	});
 
 	const onLoadKakaoMap = (k: string) => {
 		window.kakao?.maps?.load(() => {
@@ -78,35 +74,74 @@ const PlaceInput = () => {
 
 	const handleClickSelect = (placeInfo: KakaoResult) => {
 		const districtArr = placeInfo.road_address_name.split(" ");
-
-		bCodeSearch({ query: placeInfo.road_address_name }).then((res) => {
-			if (res.data) {
-				setRequestInputs({
-					...requestInputs,
-					place: {
-						place: placeInfo.place_name,
-						districts: {
-							code: res.data.documents[0].address.b_code,
-							name: `${districtArr[0]} ${districtArr[1]}`,
-						},
-						address: placeInfo.road_address_name,
-					},
-				});
-				setKeyword("");
-				setSearchOpen(false);
-			}
+		setRequestInputs({
+			...requestInputs,
+			place: {
+				place: placeInfo.place_name,
+				districts: {
+					code: "", // 중복확인 후 법정동 코드 저장
+					name: `${districtArr[0]} ${districtArr[1]}`,
+				},
+				address: placeInfo.road_address_name,
+			},
 		});
+
+		setKeyword("");
+		setSearchOpen(false);
 	};
+
+	const handleChangeCustomPlace = (
+		e: React.ChangeEvent<HTMLInputElement>,
+		id: string
+	) => {
+		if (id === "place" || id === "address1" || id === "address2") {
+			setCustomPlace({
+				...customPlace,
+				[id]: e.target.value,
+			});
+		}
+	};
+
+	const handleDeleteCustomPlace = (e: React.MouseEvent, id: string) => {
+		if (id === "place" || id === "address1" || id === "address2") {
+			setCustomPlace({
+				...customPlace,
+				[id]: "",
+			});
+		}
+	};
+
+	useEffect(() => {
+		if (isInputOpen) {
+			const districtArr = customPlace.address1.split(" ");
+			setRequestInputs({
+				...requestInputs,
+				place: {
+					place: customPlace.place,
+					districts: {
+						code: "", // 중복확인 후 법정동 코드 저장
+						name: `${districtArr[0]} ${districtArr[1]}`,
+					},
+					address:
+						customPlace.address1 +
+						(customPlace.address2 && `${customPlace.address2}`),
+				},
+			});
+		}
+	}, [customPlace]);
 
 	return (
 		<StyledPlaceInput>
-			<SearchInput
-				value={requestInputs.place.place}
-				handleClickSearchBtn={() => setSearchOpen(!isSearchOpen)}
-				id="place"
-				placeholder="장소이름"
-				label="장소 *"
-			/>
+			{!isInputOpen && (
+				<SearchInput
+					value={requestInputs.place.place}
+					handleClickInput={() => setSearchOpen(!isSearchOpen)}
+					handleClickSearchBtn={() => setSearchOpen(!isSearchOpen)}
+					id="place"
+					placeholder="장소이름"
+					label="장소 *"
+				/>
+			)}
 			{isSearchOpen && (
 				<StyledSearchListContainer>
 					<div className="inputContainer">
@@ -136,18 +171,72 @@ const PlaceInput = () => {
 									</button>
 								</li>
 							))}
+							<li>
+								<div>
+									<p>직접 입력하기</p>
+								</div>
+								<button
+									type="button"
+									onClick={() => {
+										setSearchOpen(false);
+										setInputOpen(true);
+										setCustomPlace({
+											place: keyword,
+											address1: "",
+											address2: "",
+										});
+									}}
+								>
+									선택
+								</button>
+							</li>
 						</StyledSearchList>
 					)}
 				</StyledSearchListContainer>
 			)}
-			<SearchInput
-				value={requestInputs.place.address}
-				id="address"
-				placeholder="주소"
-				label=""
-				hideLabel
-				hideButton
-			/>
+			{isInputOpen && (
+				<StyledPlaceCustomInput>
+					<BasicInput
+						label="장소 *"
+						value={customPlace.place}
+						id="place"
+						placeholder="장소이름"
+						handleInputChange={(e) => handleChangeCustomPlace(e, "place")}
+						handleInputDelete={(e) => handleDeleteCustomPlace(e, "place")}
+					/>
+
+					<div className="address">
+						<BasicInput
+							label=""
+							hideLabel
+							value={customPlace.address1}
+							id="address1"
+							placeholder="주소 (도로명, 지번)"
+							handleInputChange={(e) => handleChangeCustomPlace(e, "address1")}
+							handleInputDelete={(e) => handleDeleteCustomPlace(e, "address1")}
+						/>
+						<BasicInput
+							label=""
+							hideLabel
+							value={customPlace.address2}
+							id="address2"
+							placeholder="상세주소 (층 수, 생략 가능)"
+							handleInputChange={(e) => handleChangeCustomPlace(e, "address2")}
+							handleInputDelete={(e) => handleDeleteCustomPlace(e, "address2")}
+						/>
+					</div>
+				</StyledPlaceCustomInput>
+			)}
+			{!isInputOpen && (
+				<SearchInput
+					value={requestInputs.place.address}
+					id="address"
+					placeholder="주소"
+					label=""
+					hideLabel
+					hideButton
+				/>
+			)}
 		</StyledPlaceInput>
 	);
 };
