+++ public/sitemap-0.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://www.o-cup.kr</loc><lastmod>2022-11-01T06:29:26.912Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.o-cup.kr/request</loc><lastmod>2022-11-01T06:29:26.913Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.o-cup.kr/search</loc><lastmod>2022-11-01T06:29:26.913Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.o-cup.kr</loc><lastmod>2022-11-13T07:43:48.897Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.o-cup.kr/request</loc><lastmod>2022-11-13T07:43:48.897Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.o-cup.kr/search</loc><lastmod>2022-11-13T07:43:48.897Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
